+++ .yo-rc.json
@@ -6,12 +6,12 @@
     "nodeVersion": "22.14.0",
     "sdksVersions": {},
     "version": "1.22.1",
-    "libraryName": "context-hub",
+    "libraryName": "BookmarkHub",
     "libraryId": "556b1e36-2a92-4088-a3a9-49b9b92c912b",
     "environment": "spo",
     "packageManager": "npm",
-    "solutionName": "context-hub",
-    "solutionShortDescription": "context-hub description",
+    "solutionName": "BookmarkHub",
+    "solutionShortDescription": "bookmark-hub description",
     "skipFeatureDeployment": true,
     "isDomainIsolated": false,
     "componentType": "webpart"
