+++ config/config.json
@@ -2,17 +2,17 @@
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/config.2.0.schema.json",
   "version": "2.0",
   "bundles": {
-    "context-hub-web-part": {
+    "bookmark-hub-web-part": {
       "components": [
         {
-          "entrypoint": "./lib/webparts/contextHub/ContextHubWebPart.js",
-          "manifest": "./src/webparts/contextHub/ContextHubWebPart.manifest.json"
+          "entrypoint": "./lib/webparts/bookmarkHub/BookmarkHubWebPart.js",
+          "manifest": "./src/webparts/bookmarkHub/BookmarkHubWebPart.manifest.json"
         }
       ]
     }
   },
   "externals": {},
   "localizedResources": {
-    "ContextHubWebPartStrings": "lib/webparts/contextHub/loc/{locale}.js"
+    "BookmarkHubWebPartStrings": "lib/webparts/bookmarkHub/loc/{locale}.js"
   }
 }
