+++ config/package-solution.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/package-solution.schema.json",
   "solution": {
-    "name": "context-hub-client-side-solution",
+    "name": "BookmarkHub-client-side-solution",
     "id": "556b1e36-2a92-4088-a3a9-49b9b92c912b",
     "version": "1.0.0.0",
     "includeClientSideAssets": true,
@@ -15,17 +15,17 @@
       "mpnId": "Undefined-1.22.1"
     },
     "webApiPermissionRequests": [
-        {
-          "resource": "Microsoft Graph",
-          "scope": "Sites.Read.All"
-        }
+      {
+        "resource": "Microsoft Graph",
+        "scope": "Sites.Read.All"
+      }
     ],
     "metadata": {
       "shortDescription": {
-        "default": "context-hub description"
+        "default": "bookmark-hub description"
       },
       "longDescription": {
-        "default": "context-hub description"
+        "default": "bookmark-hub description"
       },
       "screenshotPaths": [],
       "videoUrl": "",
@@ -33,14 +33,14 @@
     },
     "features": [
       {
-        "title": "context-hub Feature",
-        "description": "The feature that activates elements of the context-hub solution.",
+        "title": "bookmark-hub Feature",
+        "description": "The feature that activates elements of the bookmark-hub solution.",
         "id": "d0710c92-f0b4-4caa-8177-173b17559866",
         "version": "1.0.0.0"
       }
     ]
   },
   "paths": {
-    "zippedPackage": "solution/context-hub.sppkg"
+    "zippedPackage": "solution/BookmarkHub.sppkg"
   }
 }
