+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "context-hub",
+  "name": "BookmarkHub",
   "version": "0.0.1",
   "private": true,
   "engines": {
