+++ src/webparts/bookmarkHub/BookmarkHubWebPart.manifest.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/spfx/client-side-web-part-manifest.schema.json",
   "id": "6c640ded-b242-40bb-8700-d936144f796a",
-  "alias": "ContextHubWebPart",
+  "alias": "BookmarkHubWebPart",
   "componentType": "WebPart",
   "version": "*",
   "manifestVersion": 2,
@@ -12,11 +12,11 @@
   "preconfiguredEntries": [{
     "groupId": "5c03119e-3074-46fd-976b-c60198311f70",
     "group": { "default": "Advanced" },
-    "title": { "default": "ContextHub" },
+    "title": { "default": "BookmarkHub" },
     "description": { "default": "A unified hub that brings together all your favorites, bookmarks, and important content from across Microsoft 365 into one organized, context‑rich view." },
     "officeFabricIconFontName": "Bookmarks",
     "properties": {
-      "description": "ContextHub"
+      "description": "BookmarkHub"
     }
   }]
 }
