+++ src/webparts/bookmarkHub/components/BookmarkHub.module.scss
@@ -1,6 +1,6 @@
 @import '~@fluentui/react/dist/sass/References.scss';
 
-.contextHub {
+.bookmarkHub {
   padding: 1em;
   color: "[theme:bodyText, default: #323130]";
 }
