+++ src/webparts/bookmarkHub/components/IBookmarkHubProps.ts
@@ -1,0 +1,7 @@
+import { BookmarkHubService } from "../../../services/BookmarkHubService";
+
+export interface IBookmarkHubProps {
+  isDarkTheme: boolean;
+  hasTeamsContext: boolean;
+  bookmarkHubService: BookmarkHubService;
+}
