--- src/webparts/contextHub/components/IContextHubProps.ts
@@ -1,7 +1,0 @@
-import { ContextHubService } from "../../../services/ContextHubService";
-
-export interface IContextHubProps {
-  isDarkTheme: boolean;
-  hasTeamsContext: boolean;
-  contextHubService: ContextHubService;
-}
